requirements.txt

pydantic
langchain
langchain-community
langchain-google-genai
langgraph
python-dotenv

fastapi==0.111.1
uvicorn[standard]==0.23.2
jinja2==3.1.2